btc_trade/requirements.txt
django
djangorestframework
psycopg2-binary
pytest
pytest-django
django_extensions
